refactor: refactor amount handling to use int64 type
- Change the type of the Amount field from float64 to int64 in multiple files
- Update the GetAmount method to return int64 instead of float64
- Modify the totalAmount calculation to handle int64 directly instead of converting to float64
- Implement error handling for fetching start and end blocks in the transaction repository
- Add context for debugging transaction fetching in the repository implementation

+++ internal/domain/core/app/query/transaction.go
@@ -66,7 +66,7 @@ func (s *TransactionQueryService) GetTotalSwapAmount(c context.Context, address,
 	for _, tx := range transactions {
 		for _, task := range campaign.Tasks {
 			if task.Criteria.PoolId == tx.ToAddress {
-				totalAmount += tx.Amount
+				totalAmount += float64(tx.Amount)
 			}
 		}
 	}
